backend/requirements.txt
fastapi
uvicorn
google-generativeai
pymupdf
requests
python-multipart
python-dotenv
msgpack
